# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Tournament system for game theory competitions",
        "title": "TJudge API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/tournaments/{id}/active-game": {
            "get": {
                "description": "Возвращает текущую активную игру турнира (null если нет)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Активная игра турнира",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tournament ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.TournamentGameWithDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Устанавливает активную игру для турнира (только для админов)",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Установить активную игру",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tournament ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ID игры",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SetActiveGameRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Активная игра установлена"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/tournaments/{id}/games/status": {
            "get": {
                "description": "Возвращает игры турнира с информацией о раундах и авто-раунде",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Статус игр турнира",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tournament ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_api_handlers.TournamentGameWithDetails"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/tournaments/{id}/games/{gameId}/leaderboard": {
            "get": {
                "description": "Возвращает таблицу лидеров для конкретной игры в турнире",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Рейтинг по игре",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tournament ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Game ID",
                        "name": "gameId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Лимит записей",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_bmstu-itstech_tjudge_internal_domain.LeaderboardEntry"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tournaments/{id}/games/{gameId}/matches": {
            "get": {
                "description": "Возвращает матчи для конкретной игры в турнире с фильтрацией по статусу",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Матчи по игре",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tournament ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Game ID",
                        "name": "gameId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по статусу (pending, running, completed, failed, cancelled)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Лимит записей",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_bmstu-itstech_tjudge_internal_domain.Match"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tournaments/{id}/games/{gameId}/programs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает программы для конкретной игры в турнире (только для админов)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Программы по игре",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tournament ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Game ID",
                        "name": "gameId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_bmstu-itstech_tjudge_internal_domain.Program"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_bmstu-itstech_tjudge_internal_domain.LeaderboardEntry": {
            "type": "object",
            "properties": {
                "draws": {
                    "type": "integer"
                },
                "losses": {
                    "type": "integer"
                },
                "program_id": {
                    "type": "string"
                },
                "program_name": {
                    "type": "string"
                },
                "rank": {
                    "type": "integer"
                },
                "rating": {
                    "type": "integer"
                },
                "team_id": {
                    "type": "string"
                },
                "team_name": {
                    "type": "string"
                },
                "total_games": {
                    "type": "integer"
                },
                "wins": {
                    "type": "integer"
                }
            }
        },
        "github_com_bmstu-itstech_tjudge_internal_domain.Match": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "error_message": {
                    "type": "string"
                },
                "game_type": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "priority": {
                    "$ref": "#/definitions/github_com_bmstu-itstech_tjudge_internal_domain.MatchPriority"
                },
                "program1_id": {
                    "type": "string"
                },
                "program2_id": {
                    "type": "string"
                },
                "round_number": {
                    "description": "Номер раунда для группировки",
                    "type": "integer"
                },
                "score1": {
                    "type": "integer"
                },
                "score2": {
                    "type": "integer"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_bmstu-itstech_tjudge_internal_domain.MatchStatus"
                },
                "tournament_id": {
                    "type": "string"
                },
                "winner": {
                    "type": "integer"
                }
            }
        },
        "github_com_bmstu-itstech_tjudge_internal_domain.MatchPriority": {
            "type": "string",
            "enum": [
                "high",
                "medium",
                "low"
            ],
            "x-enum-varnames": [
                "PriorityHigh",
                "PriorityMedium",
                "PriorityLow"
            ]
        },
...
        "github_com_bmstu-itstech_tjudge_internal_domain.MatchStatus": {
            "type": "string",
            "enum": [
                "pending",
                "running",
                "completed",
                "failed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "MatchPending",
                "MatchRunning",
                "MatchCompleted",
                "MatchFailed",
                "MatchCancelled"
            ]
        },
        "github_com_bmstu-itstech_tjudge_internal_domain.Program": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "game_id": {
                    "type": "string"
                },
                "game_type": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "team_id": {
                    "type": "string"
                },
                "tournament_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
        "internal_api_handlers.SetActiveGameRequest": {
            "type": "object",
            "properties": {
                "game_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_handlers.TournamentGameWithDetails": {
            "type": "object",
            "properties": {
                "auto_round_enabled": {
                    "type": "boolean"
                },
                "auto_round_interval_seconds": {
                    "type": "integer"
                },
                "auto_round_last_run_at": {
                    "type": "string"
                },
                "current_round": {
                    "type": "integer"
                },
                "game_display_name": {
                    "type": "string"
                },
                "game_id": {
                    "type": "string"
                },
                "game_name": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "round_completed": {
                    "type": "boolean"
                },
                "round_completed_at": {
                    "type": "string"
                },
                "tournament_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token (format: \"Bearer {token}\")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}